% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vegetation_functions.R
\name{alru_dom}
\alias{alru_dom}
\title{Percent of Alru landcover in the watershed
This functions takes the GIS_Stats/Vegetation/Data/alru_dom dataset
First it obtains the area in hectares --> drop_units(st_area(polygon2process)/10000) <-- for the watershed
Then it obtains zonal statistics - the number of pixels of alru_dom -->exact_extract(alru_dom.ras,polygon2process,'count') <--
Because each pixel is 30m it calculates the area of alru_dom by multiplying the count * 0.09 and then divides by the area to
obtain percentage}
\usage{
alru_dom(polygon2process, predictor_geometry, ...)
}
\arguments{
\item{polygon2process}{this is a geojson string for the watershed}

\item{...}{The geojson is converted to an object of type sf -->polygon2process<-geojson_sf(polygon2process)<--}
}
\value{
this functions returns one value which is the percentage of alru_dom
}
\description{
Percent of Alru landcover in the watershed
This functions takes the GIS_Stats/Vegetation/Data/alru_dom dataset
First it obtains the area in hectares --> drop_units(st_area(polygon2process)/10000) <-- for the watershed
Then it obtains zonal statistics - the number of pixels of alru_dom -->exact_extract(alru_dom.ras,polygon2process,'count') <--
Because each pixel is 30m it calculates the area of alru_dom by multiplying the count * 0.09 and then divides by the area to
obtain percentage
}
