% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{DOY}
\alias{DOY}
\title{get Julian day and change column name to DOY}
\usage{
DOY(JulianDate, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
get Julian day and change column name to DOY
}
