% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{east}
\alias{east}
\title{Assess if the point falls in an eastern Oregon ecoregion
database used -- GIS_Data/Metrics/Oregon/Data/OR_EastWest_Eco
The function first makes sure that the ecoregion is in the same CRS as the points
Then in intersects the point with the ecoregions layer and creates a column named "east"
this new column is filled with 0's. The functions then assesses if the points have the
attribute "East" in the "EastWest" column. If they do, then assigns a value of 1, else 0}
\usage{
east(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value 1 or 0 - 1 if the point is the eastern Oregon ecoregion, else 0
}
\description{
Assess if the point falls in an eastern Oregon ecoregion
database used -- GIS_Data/Metrics/Oregon/Data/OR_EastWest_Eco
The function first makes sure that the ecoregion is in the same CRS as the points
Then in intersects the point with the ecoregions layer and creates a column named "east"
this new column is filled with 0's. The functions then assesses if the points have the
attribute "East" in the "EastWest" column. If they do, then assigns a value of 1, else 0
}
