% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{ECO3}
\alias{ECO3}
\title{Eco Region level 3 (number) of the point
database used GIS_Stats/Ecoregion/Data/Eco_Level_III_US.shp
The functions first makes sure that only one column "US_L3CODE" is present in the attribute table
by creating a subset myvars <- "US_L3CODE" / Eco3_PT.vec <- Eco3_PT.vec\link{myvars}
Then it transforms the Eco_Level_III_US.shp so that it shares the same CRS with the points
it then intersects the point with the Eco_Level_III_US.shp layer and just pulls the value for the
"US_L3CODE" attribute}
\usage{
ECO3(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value: the ecoregion level 3 for the point
}
\description{
Eco Region level 3 (number) of the point
database used GIS_Stats/Ecoregion/Data/Eco_Level_III_US.shp
The functions first makes sure that only one column "US_L3CODE" is present in the attribute table
by creating a subset myvars <- "US_L3CODE" / Eco3_PT.vec <- Eco3_PT.vec\link{myvars}
Then it transforms the Eco_Level_III_US.shp so that it shares the same CRS with the points
it then intersects the point with the Eco_Level_III_US.shp layer and just pulls the value for the
"US_L3CODE" attribute
}
