% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{ECO4}
\alias{ECO4}
\title{Eco Region level 4 of the point
database used GIS_Stats/Ecoregion/Data/us_eco_l4_no_st.shp
The functions first makes sure that only one column "US_L4CODE" is present in the attribute table
by creating a subset myvars <- "US_L4CODE" / Eco4_PT.vec\link{myvars}
Then it transforms the us_eco_l4_no_st.shp so that it shares the same CRS with the points
it then intersects the point with the Eco_Level_III_US.shp layer and just pulls the value for the
"US_L4CODE" attribute}
\usage{
ECO4(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value: the ecoregion level 4 value for the point
}
\description{
Eco Region level 4 of the point
database used GIS_Stats/Ecoregion/Data/us_eco_l4_no_st.shp
The functions first makes sure that only one column "US_L4CODE" is present in the attribute table
by creating a subset myvars <- "US_L4CODE" / Eco4_PT.vec\link{myvars}
Then it transforms the us_eco_l4_no_st.shp so that it shares the same CRS with the points
it then intersects the point with the Eco_Level_III_US.shp layer and just pulls the value for the
"US_L4CODE" attribute
}
