% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation_functions.R
\name{ELEV_RANGE}
\alias{ELEV_RANGE}
\title{Range between max and min elevations
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size.
The function first obtains the max elevation in the watershed, and then the minimum and finally
it obtains the difference between the two values}
\usage{
ELEV_RANGE(polygon2process, USGS_NED, ...)
}
\arguments{
\item{polygon2process}{}
}
\value{
a single value, the elevation range in the watershed
}
\description{
Range between max and min elevations
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size.
The function first obtains the max elevation in the watershed, and then the minimum and finally
it obtains the difference between the two values
}
