% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation_functions.R
\name{ELEV_SITE_CV}
\alias{ELEV_SITE_CV}
\title{Elevation coefficient of variation at the point
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size.
The function first makes sure that the point has a CRS information in meters
because the st_buffer function does not work well with decimal degrees
it then creates a area of influence AOI around the point of 150 meters (buffer).
Extracts the mean and the standard deviation of the elevation within this AOI, and then
it calculates the coefficient of variation}
\usage{
ELEV_SITE_CV(points2process, USGS_NED, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value, the coefficient of variation of elevation in an 150 m buffer around the point
}
\description{
Elevation coefficient of variation at the point
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size.
The function first makes sure that the point has a CRS information in meters
because the st_buffer function does not work well with decimal degrees
it then creates a area of influence AOI around the point of 150 meters (buffer).
Extracts the mean and the standard deviation of the elevation within this AOI, and then
it calculates the coefficient of variation
}
