% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation_functions.R
\name{ELEV_SITE_SQRT}
\alias{ELEV_SITE_SQRT}
\title{Square root of elevation at the point
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size.
the function first extracts the elevation at the point. This elevation is then divided by
10 and then the square root is extracted}
\usage{
ELEV_SITE_SQRT(points2process, USGS_NED, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value which is the square root of the elevation at the point
}
\description{
Square root of elevation at the point
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size.
the function first extracts the elevation at the point. This elevation is then divided by
10 and then the square root is extracted
}
