% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation_functions.R
\name{ELVmean_WS}
\alias{ELVmean_WS}
\title{Mean elevation across the watershed
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size}
\usage{
ELVmean_WS(polygon2process, USGS_NED, ...)
}
\arguments{
\item{polygon2process}{}
}
\value{
a single value the mean elevation value for the watershed
}
\description{
Mean elevation across the watershed
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size
}
