% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation_functions.R
\name{ELVmean_WS_100}
\alias{ELVmean_WS_100}
\title{Watershed mean elevation divided by 100}
\usage{
ELVmean_WS_100(polygon2process, USGS_NED, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Watershed mean elevation divided by 100
}
