% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation_functions.R
\name{ELVmin_WS}
\alias{ELVmin_WS}
\title{Average of min elevation in the watershed
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size}
\usage{
ELVmin_WS(polygon2process, USGS_NED, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value the minimum elevation value for the watershed
}
\description{
Average of min elevation in the watershed
The function requires that a Google Earth Engine GEE object be created
USGS_NED National Elevation Dataset. It uses rgee rgee::ee_extract function to conduct
zonal statistics. The resolution (pixel size to use) can be changed if desired by
modifying scale=. Now it is using a 90x90 m pixel size
}
