% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{ER13}
\alias{ER13}
\title{The level 3 ecoregion number 23
databased used GIS_Stats/Ecoregion/Data/Eco_Level_III_US.shp
The point is first transformed to a CRS in meters
The functions then makes sure that only one column "US_L3CODE" is present in the attribute table
by creating a subset myvars <- "US_L3CODE" / Eco3_PT.vec\link{myvars}}
\usage{
ER13(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value "Y" if the ecoregion of the point is 23, "N" otherwise
}
\description{
Then it intersects the point with the Eco_Level_III_US.shp layer and just pulls the value for the
"US_L3CODE" attribute. A new column "ER13" is created whereby if the intersected value is 23 then it
will be populated with "Y", else it will be populated with "N"
}
