% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vegetation_functions.R
\name{Evergr_ave}
\alias{Evergr_ave}
\title{Percent of Evergreen landcover in the watershed
dataset used GIS_Stats/Vegetation/Data/evergr
The function first obtains the watershed area in hectares drop_units(st_area(polygon2process)/10000)
It then uses the exact_extract function to obtain zonal statistics for the raster, the sum of pixels within the watershed
The sum is obtained because all the pixels have a value of 1. Then in converts the sum of pixels to area
by multiplying 0.09 (30 x 30 m pixels equal 900 m2 equal 0.09 hectares). This value is then divided by the
area of the watershed and the percent (in decimal units is extracted)}
\usage{
Evergr_ave(polygon2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{
a single value of percent of evergreen landcover in the watershed in decimal values
}
\description{
Percent of Evergreen landcover in the watershed
dataset used GIS_Stats/Vegetation/Data/evergr
The function first obtains the watershed area in hectares drop_units(st_area(polygon2process)/10000)
It then uses the exact_extract function to obtain zonal statistics for the raster, the sum of pixels within the watershed
The sum is obtained because all the pixels have a value of 1. Then in converts the sum of pixels to area
by multiplying 0.09 (30 x 30 m pixels equal 900 m2 equal 0.09 hectares). This value is then divided by the
area of the watershed and the percent (in decimal units is extracted)
}
