% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{extract_point_value}
\alias{extract_point_value}
\title{simple point value extraction, raster used is specified in the database under the geometry_file_path}
\usage{
extract_point_value(point2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
simple point value extraction, raster used is specified in the database under the geometry_file_path
}
