% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions.R
\name{extract_watershed_mean}
\alias{extract_watershed_mean}
\title{simple watershed mean, raster used is specified in the database under the geometry_file_path}
\usage{
extract_watershed_mean(
  polygon2process,
  predictor_name,
  predictor_geometry,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
simple watershed mean, raster used is specified in the database under the geometry_file_path
}
