% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrology_functions.R
\name{GW_P_Sp_Mx}
\alias{GW_P_Sp_Mx}
\title{Ground Water Index across the watershed}
\usage{
GW_P_Sp_Mx(polygon2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Ground Water Index across the watershed
}
