% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{HV_UPPERPLATTE}
\alias{HV_UPPERPLATTE}
\title{Is the point within the High Valleys and Upper North Platte ecoregion, WY}
\usage{
HV_UPPERPLATTE(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Is the point within the High Valleys and Upper North Platte ecoregion, WY
}
