% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamCat_functions.R
\name{inLOOP}
\alias{inLOOP}
\title{concatenate list objects into an "IN" string for insertion into SQLite queries (needed for StreamCat)}
\usage{
inLOOP(inSTR, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
concatenate list objects into an "IN" string for insertion into SQLite queries (needed for StreamCat)
}
