% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_functions.R
\name{lat}
\alias{lat}
\title{Latitude of the point
This function returns the Y coordinate of the point in decimal degrees
The st_coordinates function returns the second column \link{,2} which is the latitude}
\usage{
lat(points2process, ...)
}
\arguments{
\item{...}{}
}
\value{
this functions returns one value which is the latitude of the point
}
\description{
Latitude of the point
This function returns the Y coordinate of the point in decimal degrees
The st_coordinates function returns the second column \link{,2} which is the latitude
}
