% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation_functions.R
\name{LOG_PRECIP_SITE}
\alias{LOG_PRECIP_SITE}
\title{Log 10 precipitation at the point}
\usage{
LOG_PRECIP_SITE(points2process, predictor_geometry, ...)
}
\arguments{
\item{predictor_geometry}{}
}
\value{

}
\description{
can be mean log precipitation at the point too if input raster is a mean
}
