% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_functions.R
\name{LOG_WSA_SQKM}
\alias{LOG_WSA_SQKM}
\title{Log watershed area in sq km}
\usage{
LOG_WSA_SQKM(polygon2process, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Log watershed area in sq km
}
