% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger}
\alias{Logger}
\title{Logger}
\value{
a \code{Logger} class (R6 class)
}
\description{
R6 class for handling NAMC oAuth2
}
\examples{


----------------------------------------------------------------
---------------------------------------------------------------
}
\section{Super class}{
\code{\link[NAMCGeopredictors:BaseClass_R6]{NAMCGeopredictors::BaseClass_R6}} -> \code{Logger}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Logger$new()}}
\item \href{#method-toggleState}{\code{Logger$toggleState()}}
\item \href{#method-getFullFilePath}{\code{Logger$getFullFilePath()}}
\item \href{#method-startLog}{\code{Logger$startLog()}}
\item \href{#method-stopLog}{\code{Logger$stopLog()}}
\item \href{#method-clone}{\code{Logger$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="NAMCGeopredictors" data-topic="BaseClass_R6" data-id="dump">}\href{../../NAMCGeopredictors/html/BaseClass_R6.html#method-dump}{\code{NAMCGeopredictors::BaseClass_R6$dump()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$new(argList = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toggleState"></a>}}
\if{latex}{\out{\hypertarget{method-toggleState}{}}}
\subsection{Method \code{toggleState()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$toggleState()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFullFilePath"></a>}}
\if{latex}{\out{\hypertarget{method-getFullFilePath}{}}}
\subsection{Method \code{getFullFilePath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$getFullFilePath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-startLog"></a>}}
\if{latex}{\out{\hypertarget{method-startLog}{}}}
\subsection{Method \code{startLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$startLog(forceLog = FALSE, suffix = "", prefix = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stopLog"></a>}}
\if{latex}{\out{\hypertarget{method-stopLog}{}}}
\subsection{Method \code{stopLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$stopLog()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
