% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamCat_functions.R
\name{MAST_mean08091314}
\alias{MAST_mean08091314}
\title{Average mean annual stream temp across 2018-2014}
\usage{
MAST_mean08091314(SQLite_file_path, COMIDs = def_sites$COMID, ...)
}
\arguments{
\item{COMIDs}{}
}
\value{

}
\description{
Average mean annual stream temp across 2018-2014
}
