% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{MRE}
\alias{MRE}
\title{Is the point within the Black hills ecoregion, WY}
\usage{
MRE(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Is the point within the Black hills ecoregion, WY
}
