% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamCat_functions.R
\name{MSST_mean08091314}
\alias{MSST_mean08091314}
\title{Average mean summer stream temp across 2018-2014}
\usage{
MSST_mean08091314(SQLite_file_path, COMIDs = def_sites$COMID, ...)
}
\arguments{
\item{COMIDs}{}
}
\value{

}
\description{
Average mean summer stream temp across 2018-2014
}
