% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamCat_functions.R
\name{MWST_mean08091314}
\alias{MWST_mean08091314}
\title{Average mean winter stream temp across 2018-2014}
\usage{
MWST_mean08091314(SQLite_file_path, COMMIDs = def_sites$COMID, ...)
}
\arguments{
\item{COMIDs}{}
}
\value{

}
\description{
Average mean winter stream temp across 2018-2014
}
