% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope_functions.R
\name{NHDSLOPE}
\alias{NHDSLOPE}
\title{NHD Plus slope value taken from the nearest stream segment
st_transform changes the CRS of the point to meters using 5070 Albers Contiguous US, now we have the point in meter (projection)
Using the new object for the point in meters, a well-known text (WKT) string will be created to query the required vector predictor
this WKT can be used as an argument in st_read to query a big vector shapefile or geopackages and just bring into memory the AOI
i.e. like a bounding without overwhelming R
Buffer the point by 200m, interest with NHD streams, extract SLOPE value
Jennifer's notes- maxdist=500 meters needs reexamined. The original python code used 200 meters. really we should be using COMID and joining to that!!}
\usage{
NHDSLOPE(points2process, geometry_input_path, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
NHD Plus slope value taken from the nearest stream segment
st_transform changes the CRS of the point to meters using 5070 Albers Contiguous US, now we have the point in meter (projection)
Using the new object for the point in meters, a well-known text (WKT) string will be created to query the required vector predictor
this WKT can be used as an argument in st_read to query a big vector shapefile or geopackages and just bring into memory the AOI
i.e. like a bounding without overwhelming R
Buffer the point by 200m, interest with NHD streams, extract SLOPE value
Jennifer's notes- maxdist=500 meters needs reexamined. The original python code used 200 meters. really we should be using COMID and joining to that!!
}
