% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geology_functions.R
\name{Pct_Alfi}
\alias{Pct_Alfi}
\title{Percent alfi soils in the watershed}
\usage{
Pct_Alfi(polygon2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Percent alfi soils in the watershed
}
