% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geology_functions.R
\name{PCT_SEDIM}
\alias{PCT_SEDIM}
\title{Percentage of watershed that is Sedimentary geology type (geology type number 5)}
\usage{
PCT_SEDIM(polygon2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Percentage of watershed that is Sedimentary geology type (geology type number 5)
}
