% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation_functions.R
\name{PPT_2MoAvg}
\alias{PPT_2MoAvg}
\title{PRISM 2 month prior moving average precipitation for the watershed}
\usage{
PPT_2MoAvg(polygon2process, CurrentYear, JulianDate, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
PRISM 2 month prior moving average precipitation for the watershed
}
