% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation_functions.R
\name{PPT_ACCUM}
\alias{PPT_ACCUM}
\title{PRISM prior year (May-April) cumulative precipitation at the point}
\usage{
PPT_ACCUM(points2process, CurrentYear, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
PRISM prior year (May-April) cumulative precipitation at the point
}
