% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation_functions.R
\name{PRCPSHORTWS}
\alias{PRCPSHORTWS}
\title{PRISM calendar year average precipitation at the watershed (CO OE/ old MMI)}
\usage{
PRCPSHORTWS(points2process, CurrentYear, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
PRISM calendar year average precipitation at the watershed (CO OE/ old MMI)
}
