% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation_functions.R
\name{precip}
\alias{precip}
\title{Precipitation from vector layer (OR- WCCP model)}
\usage{
precip(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Precipitation from vector layer (OR- WCCP model)
}
