% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_predictors.R
\name{process_box_predictors}
\alias{process_box_predictors}
\title{predictors for each box}
\usage{
process_box_predictors(boxId, config = config)

process_box_predictors(boxId, config = config)
}
\arguments{
\item{boxId}{}

\item{config}{}
}
\value{


none
}
\description{
predictors for each box

Process box predictor
}
\details{
saving each predictor for each sample one at a time in the database
}
