% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_predictors.R
\name{process_predictors}
\alias{process_predictors}
\title{run all predictors for all samples at once}
\usage{
process_predictors()
}
\value{

}
\description{
run all predictors for all samples at once
}
