% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{SFLR}
\alias{SFLR}
\title{Is the point within the Southern Foothills and Laramie Range ecoregion, WY}
\usage{
SFLR(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Is the point within the Southern Foothills and Laramie Range ecoregion, WY
}
