% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope_functions.R
\name{Slope_WS}
\alias{Slope_WS}
\title{Watershed slope using flow length- NV MMI model}
\usage{
Slope_WS(polygon2process, geometry_input_path, USGS_NED, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
total elevation change in the watershed/ arcgis flow length
}
