% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope_functions.R
\name{slpavg}
\alias{slpavg}
\title{Average Watershed slope}
\usage{
slpavg(polygon2process, USGS_NED, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
slope of each cell in a DEM and then averaged across the watershed
}
