% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elevation_functions.R
\name{SQRT_TOPO}
\alias{SQRT_TOPO}
\title{Square root of topo from the line shapefile (OLD CO OE or MMI)}
\usage{
SQRT_TOPO(points2process, predictor_geometry, geometry_input_path, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Square root of topo from the line shapefile (OLD CO OE or MMI)
}
