% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoregion_functions.R
\name{SR_BIGHORNS}
\alias{SR_BIGHORNS}
\title{Is the point within the Southern Rockies and Bighorn Mountains ecoregion, WY}
\usage{
SR_BIGHORNS(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Is the point within the Southern Rockies and Bighorn Mountains ecoregion, WY
}
