% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamCat_functions.R
\name{StreamCat_all}
\alias{StreamCat_all}
\title{Get all StreamCat predictors from SQLite database on S3 plus avg temps}
\usage{
StreamCat_all(SQLite_file_path, COMIDs, ...)
}
\arguments{
\item{COMIDs}{}
}
\value{

}
\description{
Get all StreamCat predictors from SQLite database on S3 plus avg temps
}
