% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamCat_functions.R
\name{StreamCat_single_pred}
\alias{StreamCat_single_pred}
\title{Get single StreamCat predictor from SQLite database on S3 to put into database}
\usage{
StreamCat_single_pred(SQLite_file_path, predictor_name, COMIDs, ...)
}
\arguments{
\item{predictor_name}{predictor abbreviation as it is in the database}

\item{COMIDs}{}
}
\value{

}
\description{
Get single StreamCat predictor from SQLite database on S3 to put into database
}
