% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_functions.R
\name{SUMMER}
\alias{SUMMER}
\title{Summer stream temp at the point (CO MMI specific)}
\usage{
SUMMER(points2process, predictor_geometry, geometry_input_path, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Summer stream temp at the point (CO MMI specific)
}
