% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_functions.R
\name{temp}
\alias{temp}
\title{Temperature at the point (Oregon specific vector layer)}
\usage{
temp(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Temperature at the point (Oregon specific vector layer)
}
