% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_functions.R
\name{Tmax_PT}
\alias{Tmax_PT}
\title{Temperature max at the point (AND divide by 10)}
\usage{
Tmax_PT(points2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Temperature max at the point (AND divide by 10)
}
