% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrology_functions.R
\name{Wb_mx_area}
\alias{Wb_mx_area}
\title{Area of the largest waterbody in the watershed}
\usage{
Wb_mx_area(polygon2process, predictor_geometry, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Area of the largest waterbody in the watershed
}
