% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature_functions.R
\name{WINTER}
\alias{WINTER}
\title{Winter stream temp at the point (CO MMI specific but only needed for old model/ OE)}
\usage{
WINTER(points2process, predictor_geometry, geometry_input_path, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Winter stream temp at the point (CO MMI specific but only needed for old model/ OE)
}
