% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_predictors.R
\name{process_box_predictors}
\alias{process_box_predictors}
\title{predictors for each box}
\usage{
process_box_predictors(boxId)
}
\arguments{
\item{boxId}{}
}
\value{

}
\description{
predictors for each box
}
