% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_predictors.R
\name{process_sample_predictors}
\alias{process_sample_predictors}
\title{Process sample predictors}
\usage{
process_sample_predictors(sampleId, config = config)
}
\arguments{
\item{config}{}
}
\value{
none
}
\description{
Process sample predictors
}
\details{
saving each predictor for each sample one at a time in the database
}
